pull Env out of Generate.hs
