+++ src/Language/Haskell/Liquid/Constraint/Axioms.hs
@@ -95,7 +95,8 @@ import Language.Fixpoint.Misc
 import Language.Haskell.Liquid.Types.Literals
 import Language.Haskell.Liquid.Transforms.RefSplit
 import Control.DeepSeq
-import Language.Haskell.Liquid.Constraint.Constraint
+
+-- import Language.Haskell.Liquid.Constraint.Constraint
 import Language.Haskell.Liquid.Constraint.ProofToCore
 
 import Language.Haskell.Liquid.WiredIn (wiredSortedSyms)
@@ -503,7 +504,7 @@ grapInt _          = return 2
 
 
 -------------------------------------------------------------------------------
----------------------  Combine Proofs  ----------------------------------------
+--------------------  Combine Proofs  ----------------------------------------
 -------------------------------------------------------------------------------
 
 makeCombineType Nothing
